refactor chore data and chorelist component
