+++ src/components/Input.js
@@ -3,9 +3,6 @@ import React from "react";
 import styles from "../componentStyles/MemberForm.css";
 
 export default class Input extends React.Component {
-  constructor(props) {
-    super(props);
-  }
   componentDidUpdate(prevProps) {
     if (!prevProps.meta.active && this.props.meta.active) {
       this.input.focus();
